Update BleTransport.swift
By default Bluejay marks a connection as needing auto reconnect after a successful connection. This means that if we enter the manager for instance and then the device becomes unavailable it will forever try to connect and fail, breaking all the disconnect logic and making me, a mere human developer, suffer from anxiety trying to understand what on earth it's doing. 

It just so happens that this behaviour is well documented in the Bluejay docs as something we may want to override and, believe it or not, I do want to override it.
As I type this, and sip on my lukewarm coffee, questioning my life choices, I see that it has indeed fixed the issue I've been struggling with. I'm sure there'll be more to come, but this one... this one is squashed.